+++ llamazure/rbac/integration_test.py
@@ -1,14 +1,18 @@
 """Integration tests for roles"""
+import logging
 import os
 from typing import Any
 
 import pytest
 
-from llamazure.azrest.azrest import AzureError
+from llamazure.azrest.models import AzureError
 from llamazure.rbac.conftest import retry
 from llamazure.rbac.resources import Groups, Users
-from llamazure.rbac.roles import Permission, RoleAssignment, RoleAssignments, RoleDefinition, RoleDefinitions, RoleOps
+from llamazure.rbac.role_asn import RoleAssignment
+from llamazure.rbac.role_def import Permission, RoleDefinition
+from llamazure.rbac.roles import RoleAssignments, RoleDefinitions, RoleOps
 
+l = logging.getLogger(__name__)
 attempts = 5
 
 
@@ -30,33 +34,37 @@ class TestRoles:
 		"""Test that thing initialise"""
 
 	@pytest.mark.integration
+	@pytest.mark.admin
 	def test_all(self, rds: RoleDefinitions, ras: RoleAssignments, role_ops: RoleOps, scopes):
 		"""Test a whole cycle of things"""
+		role_name = "llamazure-rbac-0"
 
-		# try to purge role
-		retry(lambda: role_ops.delete_by_name("llamazure-rbac"), AzureError)
+		l.info(f"try to purge role name={role_name}")
+		retry(lambda: role_ops.delete_by_name(role_name), AzureError)
 
 		scope = scopes["sub0"]
 		scope_other = scopes["sub1"]
 
+		l.info(f"try to create role name={role_name}")
 		response = rds.put(
 			RoleDefinition.Properties(
-				roleName="llamazure-rbac",
+				roleName=role_name,
 				description="Test creating a role",
 				permissions=[Permission(actions=["Microsoft.Authorization/*/read"])],
 				assignableScopes=[scope, scope_other],
+				type="CustomRole",
 			),
 			scope=scope,
 		)
 
 		def assert_role_created():
-			role = rds.get_by_name("llamazure-rbac")
+			role = rds.get_by_name(role_name)
 			assert role
 			# compare the properties because that's what matters and `get_by_name` gives the root scope
 			assert role.properties == response.properties
 			return role
 
-		role = retry(assert_role_created, {KeyError})
+		role = retry(assert_role_created, {KeyError}, msg="assert_role_created")
 
 		def assert_role_assigned():
 			asn = ras.put(RoleAssignment.Properties(roleDefinitionId=role.rid, principalId="094238bf-5cf8-412e-8773-8e2a39c45616", principalType="User", scope=scope))
@@ -65,26 +73,42 @@ class TestRoles:
 
 		asn = retry(assert_role_assigned, AzureError)
 
+		# explicitly make a `put` that already exists
+		retry(assert_role_assigned, AzureError)
+
+		l.info("deleting RoleAssignment")
 		ras.DeleteById(asn.rid)
 
-		rds.delete_by_name(role.properties.roleName)
+		l.info("cleanup role")
+		retry(lambda: rds.delete_by_name(role.properties.roleName), AzureError)
 
 	@pytest.mark.integration
+	@pytest.mark.admin
 	def test_assign(self, rds: RoleDefinitions, ras: RoleAssignments, role_ops: RoleOps, me, scopes):
-		role_name = "llamazure-rbac-asn"
+		"""
+		Test that we can:
+			- create RoleDefinitions
+			- assign them
+			- assign in a scope outside the assignableScopes
+			- delete RoleDefinitions and the associated role assignments
+		"""
+
+		role_name = "llamazure-rbac-asn-0"
 		retry(lambda: role_ops.delete_by_name(role_name), AzureError)
 
 		sub0, sub1 = scopes["sub0"], scopes["sub1"]
 
 		role = rds.put(
 			RoleDefinition.Properties(
-				roleName="llamazure-rbac-asn",
+				roleName=role_name,
 				description="test finding assignments",
 				permissions=[Permission(actions=["Microsoft.Authorization/*/read"])],
+				type="CustomRole",
 			),
 			scope=sub0,
 		)
 		assert role
+		assert isinstance(role.rid, str)
 		assert role.rid.startswith(sub0)
 
 		def mk_asn(scope):
@@ -114,39 +138,46 @@ class TestRoles:
 			assignments_on_sub1 = ras.list_for_role_at_scope(role, sub1)
 			assert len(assignments_on_sub1) == 1
 
-		retry(exercise_listing_at_scope, {AssertionError, AzureError})
+		retry(exercise_listing_at_scope, {AssertionError, AzureError, KeyError})
 
 		# check removing assignments
 		ras.remove_all_assignments(role)
 
 		# cleanup
-		role_ops.delete_by_name(role_name)
+		retry(lambda: role_ops.delete_by_name(role_name), AzureError)
 
 
 class TestUsersAndGroups:
+	"""Tests for RBAC users and groups"""
+
 	@pytest.mark.integration
+	@pytest.mark.admin
 	def test_list_users(self, users: Users):
 		me = users.current()
 		all_users = users.list()
 		assert me["id"] in {e["id"] for e in all_users}, "did not find self in all users"
 
 	@pytest.mark.integration
+	@pytest.mark.admin
 	def test_list_users_with_groups(self, users: Users, me):
 		users_with_groups = users.list_with_memberOf()
 		me = next(e for e in users_with_groups if e["id"] == me["id"])
 		assert me["memberOf"]
 
 	@pytest.mark.integration
+	@pytest.mark.admin
 	def test_list_groups(self, groups: Groups):
 		all_groups = groups.list()
 		assert all_groups
 
 	@pytest.mark.integration
+	@pytest.mark.admin
 	def test_list_groups_with_members(self, groups: Groups):
 		all_groups_with_members = groups.list_with_memberships()
 		assert any("transitiveMembers" in g for g in all_groups_with_members)
 
 	@pytest.mark.integration
+	@pytest.mark.admin
 	def test_list_memberships_complete(self, users: Users, groups: Groups):
 		all_users = users.list_with_memberOf()
 		all_groups = {g["id"]: g for g in groups.list_with_memberships()}