+++ llamazure/azrest/changelog.md
@@ -1,5 +1,13 @@
 # 0
 
+# 0.1
+
+### 0.1.0
+
+- feature: completely rebuild everything
+- feature: codegen interface from Azure OpenAPI specs
+- feature: batch api
+
 # 0.0
 
 ### 0.0.2