+++ llamazure/msgraph/changelog.md
@@ -2,6 +2,10 @@
 
 # 0.0
 
+### 0.0.3
+
+- fix: correct scope for service principals
+
 ### 0.0.2
 
 - deps: update llamazure.rid